Internal

Modules

Exceptions — Module

f

Exceptions.Internal — Module

f

Constants

Types

Exceptions.Internal.DocstringIsNotAString — Type

Exception thrown when the passed expression for docstring does not yield a string.

Exceptions.Internal.ErrorMessageIsNotAString — Type

Exception thrown when the passed expression(s) for error message (do)es not yield a string.

Exceptions.Internal.OnlyOneContext — Type

Exception thrown when more than one context has been passed.

Exceptions.Internal.OnlyOneEquation — Type

Exception thrown when non-context equation has been passed.

Macros

Exceptions.Internal.@exception — Macro

@exception(macro_name::Symbol, args::Expr...; context::Expr=:()) -> Expr

Create a macro to create exceptions. Optionally inject context before defining the structure.

Arguments

macro_name::Symbol: name of the macro

args::Tuple{Vararg{Expr}}: a set of fields to be defined in the exception structure

Keywords

context::Expr=:(): expression evaluated before defining the exception structure

Returns

Expr: new macro definition

Throws

OnlyOneContext: more than one context has been passed

OnlyOneEquation: more than one equation has been passed

Example

using Exceptions
using Suppressor
using SyntaxTree

macro_name = :name
args = (:(arg1::String), :(arg2::Int))
context = :()

exceptions = Dict{Symbol, Any}()
exceptions[:DocstringIsNotAString] = Exceptions.Internal.DocstringIsNotAString
exceptions[:ErrorMessageIsNotAString] = Exceptions.Internal.ErrorMessageIsNotAString

d1 = @capture_out @macroexpand(@exception(name, arg1::String, arg2::Int)) |>
     linefilter! |> dump

d2 = @capture_out quote
    macro $(macro_name)(
        exception_name::Symbol,
        docstring::Union{Expr, String},
        error_message_bits::Union{Expr, String}...,
    )
        args = $(args)

        # Defaults
        module_name = __module__
        error_header = "$(module_name).$(exception_name):"

        $(context)

        e = $(exceptions)

        return esc(
            quote
                # Checks
                if !($(docstring) isa String)
                    throw($(e[:DocstringIsNotAString])())
                end
                if !($(error_message_bits...) isa String)
                    throw($(e[:ErrorMessageIsNotAString])())
                end

                @doc $(docstring)
                mutable struct $(exception_name) <: Exception
                    $(args...)
                    $(exception_name)($(args...)) = new($(args...))
                end

                Base.showerror(io::IO, e::$(module_name).$(exception_name)) =
                print(
                    io, string(
                        '\n', '\n',
                        $(error_header), '\n',
                        $(error_message_bits...), '\n',
                    )
                )
            end
        )
    end
end |> linefilter! |> dump

d1 == d2